web/requirements.txt
Flask
flask-restful
psycopg2-binary>=2.8
requests
pyjwt